foodgam/requirements.txt
asgiref==3.3.4
Django==3.2.3
django-extra-views==0.13.0
django-filter==2.4.0
djangorestframework==3.12.4
Pillow==8.2.0
pytz==2021.1
reportlab==3.5.67
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
